app/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
requests
beautifulsoup4
requests[socks]
redis
cryptography